# doc/contracts/backend-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: LabRat Backend API
  version: 1.0.0
  description: >-
    Contract-first API for the NestJS modular monolith. All documented
    operations have closed DTOs and authoritative v1 behavior.
    Public Guest identities are restricted to one demo project with read/export
    access. Authenticated Guest writes, AI requests and invitation redemption
    return 403 public_guest_read_only; only GET/HEAD and POST logout are allowed.
    Guest request limits may return 429 too_many_requests on authenticated reads.
servers:
  - url: http://127.0.0.1:8788
    description: Local v1 development server
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Tenancy]
      operationId: getProject
      x-labrat-capability: read
      responses:
        "200":
          description: Full or minimum authorized Project shell.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Tenancy]
      operationId: updateProject
      x-labrat-capability: propose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/projects/{projectId}/chart-style-profiles:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [Charts]
      operationId: listChartStyleProfiles
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/PageLimit" }
        - name: includeArchived
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Bounded chart-style profiles.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChartStyleProfilePage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Charts]
      operationId: createChartStyleProfile
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateChartStyleProfileRequest" }
      responses:
        "201":
          description: Chart-style profile and first immutable version created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChartStyleProfileDetail" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/chart-style-profiles/{chartStyleProfileId}:
    parameters:
      - name: chartStyleProfileId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/OpaqueId" }
    get:
      tags: [Charts]
      operationId: getChartStyleProfile
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200":
          description: Profile and immutable versions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChartStyleProfileDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/chart-style-profiles/{chartStyleProfileId}/versions:
    parameters:
      - name: chartStyleProfileId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/OpaqueId" }
    post:
      tags: [Charts]
      operationId: createChartStyleProfileVersion
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateChartStyleProfileVersionRequest" }
      responses:
        "201":
          description: Immutable profile version created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChartStyleProfileDetail" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/chart-style-profiles/{chartStyleProfileId}/archive:
    parameters:
      - name: chartStyleProfileId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/OpaqueId" }
    post:
      tags: [Charts]
      operationId: archiveChartStyleProfile
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      responses:
        "200":
          description: Profile archived without deleting version history.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChartStyleProfileResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/projects/{projectId}/reusable-chart-templates:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [Charts]
      operationId: listReusableChartTemplates
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      parameters:
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/PageLimit" }
        - name: includeArchived
          in: query
          required: false
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Bounded reusable chart templates.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplatePage" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Charts]
      operationId: createReusableChartTemplate
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReusableChartTemplateRequest" }
      responses:
        "201":
          description: Reusable chart template and first immutable version created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplateDetail" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/reusable-chart-templates/{reusableChartTemplateId}:
    parameters:
      - name: reusableChartTemplateId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/OpaqueId" }
    get:
      tags: [Charts]
      operationId: getReusableChartTemplate
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200":
          description: Template and immutable versions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplateDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/reusable-chart-templates/{reusableChartTemplateId}/versions:
    parameters:
      - name: reusableChartTemplateId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/OpaqueId" }
    post:
      tags: [Charts]
      operationId: createReusableChartTemplateVersion
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateReusableChartTemplateVersionRequest" }
      responses:
        "201":
          description: Immutable reusable template version created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplateDetail" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/reusable-chart-templates/{reusableChartTemplateId}/archive:
    parameters:
      - name: reusableChartTemplateId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/OpaqueId" }
    post:
      tags: [Charts]
      operationId: archiveReusableChartTemplate
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      responses:
        "200":
          description: Template archived without deleting version history.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplateResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/chart-specs/{chartSpecId}/template-eligibility:
    parameters: [{ $ref: "#/components/parameters/ChartSpecId" }]
    get:
      tags: [Charts]
      operationId: getChartTemplateEligibility
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200":
          description: Eligibility and bounded compatibility contract for reuse.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplateEligibility" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/reusable-chart-template-versions/{templateVersionId}/applications:
    parameters:
      - name: templateVersionId
        in: path
        required: true
        schema: { $ref: "#/components/schemas/OpaqueId" }
    post:
      tags: [Charts, Analysis]
      operationId: applyReusableChartTemplate
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplyReusableChartTemplateRequest" }
      responses:
        "200":
          description: Exact idempotent replay.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplateApplicationResponse" }
        "201":
          description: New compatibility check and optional deterministic analysis run.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ReusableChartTemplateApplicationResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/chart-specs/{chartSpecId}:
    parameters: [{ $ref: "#/components/parameters/ChartSpecId" }]
    get:
      tags: [Charts]
      operationId: getChartSpec
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/ChartSpecObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: labrat_session
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
...
    ChartSpecId:
      name: chartSpecId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 200, pattern: "^[A-Za-z0-9._:-]+$" }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 500 }
...
    PageLimit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    Unauthorized:
      description: Authentication is missing or expired.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: The actor lacks the required capability.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found in the authorized scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Version, state, stale-head or idempotency conflict.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Request failed bounded validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ChartSpecObject:
      description: One complete immutable ChartSpec.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ChartSpecResponse" }
...
  schemas:
...
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 200
    Timestamp:
      type: string
      format: date-time
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string, minLength: 1, maxLength: 100 }
            message: { type: string, minLength: 1, maxLength: 1000 }
            requestId: { type: string, minLength: 1, maxLength: 200 }
            details:
              type: object
              additionalProperties: true
...
    Project:
      type: object
      additionalProperties: false
      required: [id, labId, name, description, status]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        labId: { $ref: "#/components/schemas/OpaqueId" }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived] }
        projectProfile: { $ref: "#/components/schemas/ProjectProfile" }
        shellOnly: { type: boolean, default: false }
        capabilities:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/Capability" }
        workflowSummary: { $ref: "#/components/schemas/ProjectWorkflowSummary" }
...
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 300 }
        description: { type: string, maxLength: 5000 }
        status: { type: string, enum: [active, archived] }
    ProjectResponse:
      type: object
      additionalProperties: false
      required: [project]
      properties:
        project: { $ref: "#/components/schemas/Project" }
...
    NullableString:
      oneOf:
        - type: string
        - type: "null"
...
    NullableOpaqueId:
      oneOf:
        - { $ref: "#/components/schemas/OpaqueId" }
        - type: "null"
...
    JsonObject:
      type: object
      additionalProperties: { $ref: "#/components/schemas/JsonValue" }
...
    AnalysisThread:
      type: object
      additionalProperties: false
      required: [id, labId, projectId, schemaVersion, status, outputTarget, inputMode, originalRequest, messageCount, planRevisionIds, analysisRunIds, acceptedAnalysisResultIds, chartSpecIds, dataSnapshotIds, browserViewIds, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        labId: { $ref: "#/components/schemas/OpaqueId" }
        projectId: { $ref: "#/components/schemas/OpaqueId" }
        schemaVersion: { type: string }
        status: { type: string }
        outputTarget: { type: string, enum: [chart, experiment_browser] }
        inputMode:
          oneOf:
            - { type: string, enum: [experiment_browser, workbook] }
            - type: "null"
        originalRequest: { type: string }
        messageCount: { type: integer, minimum: 0 }
        messages: { type: array, maxItems: 200, items: { $ref: "#/components/schemas/JsonObject" } }
        planRevisionIds: { type: array, maxItems: 1000, items: { $ref: "#/components/schemas/OpaqueId" } }
        analysisRunIds: { type: array, maxItems: 1000, items: { $ref: "#/components/schemas/OpaqueId" } }
        acceptedAnalysisResultIds: { type: array, maxItems: 1000, items: { $ref: "#/components/schemas/OpaqueId" } }
        chartSpecIds: { type: array, maxItems: 1000, items: { $ref: "#/components/schemas/OpaqueId" } }
        dataSnapshotIds: { type: array, maxItems: 1000, items: { $ref: "#/components/schemas/OpaqueId" } }
        browserViewIds: { type: array, maxItems: 1000, items: { $ref: "#/components/schemas/OpaqueId" } }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
        createdBy: { $ref: "#/components/schemas/OpaqueId" }
        updatedBy: { $ref: "#/components/schemas/NullableOpaqueId" }
...
    CreateChartStyleProfileRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 1000 }
        style: { $ref: "#/components/schemas/JsonObject" }
        definition: { $ref: "#/components/schemas/JsonObject" }
    CreateChartStyleProfileVersionRequest:
      type: object
      additionalProperties: false
      properties:
        style: { $ref: "#/components/schemas/JsonObject" }
        definition: { $ref: "#/components/schemas/JsonObject" }
    ChartStyleProfileSummary:
      type: object
      additionalProperties: false
      required: [id, schemaVersion, name, description, status, currentVersionId, currentVersion, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        schemaVersion: { type: string, const: labrat.chartStyleProfile.v1 }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived] }
        currentVersionId: { $ref: "#/components/schemas/NullableOpaqueId" }
        currentVersion:
          oneOf:
            - { type: integer, minimum: 1 }
            - type: "null"
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    ChartStyleProfile:
      type: object
      additionalProperties: false
      required: [id, labId, projectId, schemaVersion, name, description, status, currentVersionId, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        labId: { $ref: "#/components/schemas/OpaqueId" }
        projectId: { $ref: "#/components/schemas/OpaqueId" }
        schemaVersion: { type: string, const: labrat.chartStyleProfile.v1 }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived] }
        currentVersionId: { $ref: "#/components/schemas/NullableOpaqueId" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
        createdBy: { $ref: "#/components/schemas/NullableOpaqueId" }
        updatedBy: { $ref: "#/components/schemas/NullableOpaqueId" }
    ChartStyleProfileVersion:
      type: object
      additionalProperties: true
      required: [id, labId, projectId, chartStyleProfileId, schemaVersion, version, status, contentHash, createdAt, createdBy, acceptedAt, acceptedBy]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        labId: { $ref: "#/components/schemas/OpaqueId" }
        projectId: { $ref: "#/components/schemas/OpaqueId" }
        chartStyleProfileId: { $ref: "#/components/schemas/OpaqueId" }
        schemaVersion: { type: string, const: labrat.chartStyleProfileVersion.v1 }
        version: { type: integer, minimum: 1 }
        status: { type: string, enum: [draft, accepted] }
        contentHash: { type: string }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        createdBy: { $ref: "#/components/schemas/NullableOpaqueId" }
        acceptedAt: { $ref: "#/components/schemas/NullableTimestamp" }
        acceptedBy: { $ref: "#/components/schemas/NullableOpaqueId" }
    ChartStyleProfilePage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ChartStyleProfileSummary" } }
        nextCursor: { $ref: "#/components/schemas/NullableString" }
    ChartStyleProfileDetail:
      type: object
      additionalProperties: false
      required: [chartStyleProfile, versions]
      properties:
        chartStyleProfile: { $ref: "#/components/schemas/ChartStyleProfile" }
        versions: { type: array, items: { $ref: "#/components/schemas/ChartStyleProfileVersion" } }
    ChartStyleProfileResponse:
      type: object
      additionalProperties: false
      required: [chartStyleProfile]
      properties:
        chartStyleProfile: { $ref: "#/components/schemas/ChartStyleProfile" }
    CreateReusableChartTemplateRequest:
      type: object
      additionalProperties: false
      required: [name, sourceChartSpecId]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 1000 }
        sourceChartSpecId: { $ref: "#/components/schemas/OpaqueId" }
        chartStyleProfileVersionId: { $ref: "#/components/schemas/NullableOpaqueId" }
    CreateReusableChartTemplateVersionRequest:
      type: object
      additionalProperties: false
      properties:
        sourceChartSpecId: { $ref: "#/components/schemas/OpaqueId" }
        chartStyleProfileVersionId: { $ref: "#/components/schemas/NullableOpaqueId" }
    ReusableChartTemplateSummary:
      type: object
      additionalProperties: false
      required: [id, schemaVersion, name, description, status, currentVersionId, currentVersion, chartType, experimentCardinality, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        schemaVersion: { type: string, const: labrat.reusableChartTemplate.v1 }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived] }
        currentVersionId: { $ref: "#/components/schemas/NullableOpaqueId" }
        currentVersion:
          oneOf:
            - { type: integer, minimum: 1 }
            - type: "null"
        chartType: { $ref: "#/components/schemas/NullableString" }
        experimentCardinality:
          oneOf:
            - { $ref: "#/components/schemas/JsonObject" }
            - type: "null"
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    ReusableChartTemplate:
      type: object
      additionalProperties: false
      required: [id, labId, projectId, schemaVersion, name, description, status, currentVersionId, createdAt, updatedAt, createdBy, updatedBy]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        labId: { $ref: "#/components/schemas/OpaqueId" }
        projectId: { $ref: "#/components/schemas/OpaqueId" }
        schemaVersion: { type: string, const: labrat.reusableChartTemplate.v1 }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived] }
        currentVersionId: { $ref: "#/components/schemas/NullableOpaqueId" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
        createdBy: { $ref: "#/components/schemas/NullableOpaqueId" }
        updatedBy: { $ref: "#/components/schemas/NullableOpaqueId" }
    ReusableChartTemplateVersion:
      type: object
      additionalProperties: true
      required: [id, labId, projectId, reusableChartTemplateId, schemaVersion, version, status, sourceChartSpecId, chartStyleProfileVersionId, contentHash, createdAt, createdBy, acceptedAt, acceptedBy]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        labId: { $ref: "#/components/schemas/OpaqueId" }
        projectId: { $ref: "#/components/schemas/OpaqueId" }
        reusableChartTemplateId: { $ref: "#/components/schemas/OpaqueId" }
        schemaVersion: { type: string, const: labrat.reusableChartTemplateVersion.v1 }
        version: { type: integer, minimum: 1 }
        status: { type: string, const: accepted }
        sourceChartSpecId: { $ref: "#/components/schemas/OpaqueId" }
        chartStyleProfileVersionId: { $ref: "#/components/schemas/NullableOpaqueId" }
        contentHash: { type: string }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        createdBy: { $ref: "#/components/schemas/NullableOpaqueId" }
        acceptedAt: { $ref: "#/components/schemas/Timestamp" }
        acceptedBy: { $ref: "#/components/schemas/NullableOpaqueId" }
    ReusableChartTemplatePage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ReusableChartTemplateSummary" } }
        nextCursor: { $ref: "#/components/schemas/NullableString" }
    ReusableChartTemplateDetail:
      type: object
      additionalProperties: false
      required: [reusableChartTemplate, versions]
      properties:
        reusableChartTemplate: { $ref: "#/components/schemas/ReusableChartTemplate" }
        versions: { type: array, items: { $ref: "#/components/schemas/ReusableChartTemplateVersion" } }
    ReusableChartTemplateResponse:
      type: object
      additionalProperties: false
      required: [reusableChartTemplate]
      properties:
        reusableChartTemplate: { $ref: "#/components/schemas/ReusableChartTemplate" }
    ReusableChartTemplateBindingRequest:
      type: object
      additionalProperties: false
      required: [slotId, columnId]
      properties:
        slotId: { $ref: "#/components/schemas/OpaqueId" }
        columnId: { $ref: "#/components/schemas/OpaqueId" }
    ApplyReusableChartTemplateRequest:
      type: object
      additionalProperties: false
      required: [experimentIds]
      properties:
        experimentIds:
          type: array
          minItems: 1
          maxItems: 500
          uniqueItems: true
          items: { $ref: "#/components/schemas/OpaqueId" }
        bindings:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/ReusableChartTemplateBindingRequest" }
    ReusableChartTemplateEligibility:
      type: object
      additionalProperties: false
      required: [schemaVersion, status, chartSpecId]
      properties:
        schemaVersion: { type: string, const: labrat.reusableChartTemplateEligibility.v1 }
        status: { type: string, enum: [eligible, ineligible] }
        chartSpecId: { $ref: "#/components/schemas/OpaqueId" }
        experimentCardinality: { $ref: "#/components/schemas/JsonObject" }
        inputSlots: { type: array, items: { $ref: "#/components/schemas/JsonObject" } }
        encoding: { $ref: "#/components/schemas/JsonObject" }
        missingDataPolicy: { $ref: "#/components/schemas/JsonObject" }
        blockers: { type: array, items: { $ref: "#/components/schemas/JsonObject" } }
    ReusableChartTemplateApplicationResponse:
      type: object
      additionalProperties: false
      required: [schemaVersion, replayed, application, compatibility, analysisThread, analysisPlanRevision, analysisRun]
      properties:
        schemaVersion: { type: string, const: labrat.reusableChartTemplateApplicationResponse.v1 }
        replayed: { type: boolean }
        application: { $ref: "#/components/schemas/JsonObject" }
        compatibility: { $ref: "#/components/schemas/JsonObject" }
        analysisThread:
          oneOf:
            - { $ref: "#/components/schemas/AnalysisThread" }
            - type: "null"
        analysisPlanRevision:
          oneOf:
            - { $ref: "#/components/schemas/AnalysisPlanRevision" }
            - type: "null"
        analysisRun:
          oneOf:
            - { $ref: "#/components/schemas/AnalysisRun" }
            - type: "null"
...